Implemented line paint so-so

+++ ov03/graphics/include/screen.h
@@ -9,7 +9,7 @@ typedef struct {
 	int _fd;
 	int width;
 	int height;
-	uint16_t *frame_buffer;
+	uint8_t *frame_buffer;
 } Screen;
 
 Screen ScreenNew( int width, int height );
